An integer clock, the float libm, and no double left in the game's arithmetic
Three changes, and after them the only two mentions of double in Blocks5/src
are EM_ASM_DOUBLE and the lookahead beside it, where a JavaScript number is an
IEEE double and nothing else will do.

The wall clock is an integer. getExactTimeUS counts microseconds since the
first call, because every caller reads the difference of two readings and a
count cannot drift at all, where a float of seconds steps by 244 us an hour in
and a double only pushes that out instead of removing it. Windows takes
QueryPerformanceCounter's int64 with the seconds and the remainder reduced
apart, so the product cannot leave the range; POSIX takes tv_sec and tv_nsec,
summed signed because the later reading's nanoseconds are regularly the
smaller. Only the browser converts from a floating-point value, and it has to:
emscripten_get_now() hands back a JavaScript number.

An unqualified sin(x) on a float was the C function. <cmath> puts the float
overloads in std:: and <math.h> puts only the double ones in the global
namespace, so the call widened to double, ran the double routine and narrowed
back - in 33 files, invisible to any search for the word "double". 164 call
sites now read sinf, cosf, floorf and their cousins, Vec::length takes sqrtf
through a VecHelper<float>, and no object in the game refers to a double libm
function any more.

The reductions are float, and one of them is gone. A texture offset still
wraps: its period is exactly representable, so fmodf divides by the right
number, and the wrap is what the shader needs rather than the CPU. A sine's
argument must not, and clockPhase is named so that nobody adds it back - sinf
reduces against the real pi itself and lands 3e-08 from the true value at any
clock, where a fmodf by a float 2*pi first divides by a constant 1.7e-07 out
and drifts with the turns discarded: 3.9e-05 a minute into a level, 2.0e-03
after an hour, 0.14 after three days. What float does cost is the product, and
that was measured: the clouds' one-texel-a-tick step is exactly 1.0000 out to
twelve hours in a single level, and Level::time starts again at every level.

Mat4 is float throughout now, gluPerspective's entries and GLU's normalize
included. It keeps the order of operations of the calls it stands for, which
costs nothing, but not their precision.

Tools/syntax.sh gained the matching gate: a double handed to a float fails the
run exactly as an integer handed to one does. It caught eight sites while this
was being written.

Nine of the nineteen oracle scenes move against the previous commit, worst 3
channel levels of 255 - sinf differing from the double routine in its last bit.
verify.py and the gate are clean over all 122 sources, both builds link, and
both smoke tests pass.

+++ Blocks5/src/bomb.cpp
@@ -98,7 +98,7 @@ void Bomb::onUpdate()
 
 						p.position = position * 16 + offset;
 						const float r = random(0.0f, 6.283f);
-						p.velocity = random(3.0f, 6.0f) * Vec2f(sin(r), cos(r));
+						p.velocity = random(3.0f, 6.0f) * Vec2f(sinf(r), cosf(r));
 						p.color = sampled + Vec4f(0.0f, 0.0f, 0.0f, random(0.3f, 0.5f));
 						p.deltaColor = Vec4f(0.0f, 0.0f, 0.0f, -p.color.a / p.lifetime);
 						p.rotation = random(0.0f, 10.0f);
@@ -182,7 +182,7 @@ void Bomb::onUpdate()
 					p.sizeOnTexture = Vec2b(16, 16);
 					p.position = position * 16 + Vec2i(8, 8);
 					const float r = random(0.0f, 6.283f);
-					p.velocity = random(13.0f, 15.0f) * Vec2f(sin(r), cos(r));
+					p.velocity = random(13.0f, 15.0f) * Vec2f(sinf(r), cosf(r));
 					p.color = Vec4f(1.0f, 1.0f, 1.0f, 0.2f);
 					p.deltaColor = Vec4f(0.0f, 0.0f, 0.0f, -0.2f / p.lifetime);
 					p.rotation = random(0.0f, 10.0f);
@@ -202,7 +202,7 @@ void Bomb::onUpdate()
 					p.sizeOnTexture = Vec2b(16, 16);
 					p.position = position * 16 + Vec2i(8, 8);
 					const float r = random(0.0f, 6.283f);
-					p.velocity = random(1.0f, 8.0f) * Vec2f(sin(r), cos(r));
+					p.velocity = random(1.0f, 8.0f) * Vec2f(sinf(r), cosf(r));
 					p.color = Vec4f(random(0.5f, 1.0f), random(0.5f, 1.0f), 1.0f, random(0.05f, 0.15f));
 					const float dc = -1.0f / (p.lifetime + random(-25, 25));
 					p.deltaColor = Vec4f(dc, dc, dc, -p.color.a / p.lifetime);
@@ -224,7 +224,7 @@ void Bomb::onUpdate()
 					p.sizeOnTexture = Vec2b(16, 16);
 					p.position = position * 16 + Vec2i(8, 8);
 					const float r = random(0.0f, 6.283f);
-					p.velocity = random(0.0f, 5.0f) * Vec2f(sin(r), cos(r));
+					p.velocity = random(0.0f, 5.0f) * Vec2f(sinf(r), cosf(r));
 					p.color = Vec4f(random(0.75f, 1.0f), random(0.4f, 0.75f), random(0.0f, 0.25f), 0.2f);
 					p.deltaColor = Vec4f(0.0f, 0.0f, 0.0f, -0.25f / p.lifetime);
 					p.rotation = random(0.0f, 10.0f);
